refactor(components): add container files for the components
