+++ client/src/containers/Lesson/Lesson.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import NavBar from '../../components/NavBar/NavBar';
-import InstructionList from '../../components/InstructionList/InstructionList';
+import InstructionList from '../../components/Lesson/InstructionList/InstructionList';
 import CodeEditor from '../../components/CodeEditor/CodeEditor';
-import TaskList from '../../components/TaskList/TaskList';
+import TaskList from '../../components/Lesson/TaskList/TaskList';
 
 export default function Landing(): JSX.Element {
   return (
