+++ client/src/styles/index.sass
@@ -4,16 +4,30 @@
 @import '../containers/App/App'
 @import '../containers/Landing/Landing'
 @import '../containers/Login/Login'
-@import '../containers/Dashboard/Dashboard'
 @import '../containers/LessonsOverview/LessonsOverview'
+@import '../containers/Dashboard/Dashboard'
 @import '../containers/Lesson/Lesson'
 @import '../containers/Error/Error'
 
+//App
 @import '../components/NavBar/NavBar'
+
+//Login
 @import '../components/LoginForm/LoginForm'
 @import '../components/RegisterForm/RegisterForm'
+
+//LessonsOverview
+@import '../components/LessonCard/LessonCard'
+
+//Dashboard
+@import '../components/OtherLesson/OtherLesson'
+@import '../components/OtherLessonList/OtherLessonList'
+@import '../components/Reward/Reward'
+@import '../components/RewardList/RewardList'
+@import '../components/UserLesson/UserLesson'
+
+//Lesson
 @import '../components/CodeEditor/CodeEditor'
 @import '../components/InstructionList/InstructionList'
 @import '../components/Instruction/Instruction'
 @import '../components/TaskList/TaskList'
-@import '../components/Task/Task'
